// web/src/api/models/DhcpLeaseView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DhcpLeaseView
 */
export interface DhcpLeaseView {
    /**
     * 
     * @type {string}
     * @memberof DhcpLeaseView
     */
    duid: string;
    /**
     * 
     * @type {number}
     * @memberof DhcpLeaseView
     */
    expire: number;
    /**
     * 
     * @type {string}
     * @memberof DhcpLeaseView
     */
    hostname: string;
    /**
     * 
     * @type {string}
     * @memberof DhcpLeaseView
     */
    hw_address: string;
    /**
     * 
     * @type {string}
     * @memberof DhcpLeaseView
     */
    ip_address: string;
    /**
     * 
     * @type {boolean}
     * @memberof DhcpLeaseView
     */
    reserved: boolean;
    /**
     * 
     * @type {string}
     * @memberof DhcpLeaseView
     */
    service: string;
    /**
     * 
     * @type {number}
     * @memberof DhcpLeaseView
     */
    state: number;
    /**
     * 
     * @type {number}
     * @memberof DhcpLeaseView
     */
    subnet_id: number;
}

/**
 * Check if a given object implements the DhcpLeaseView interface.
 */
export function instanceOfDhcpLeaseView(value: object): value is DhcpLeaseView {
    if (!('duid' in value) || value['duid'] === undefined) return false;
    if (!('expire' in value) || value['expire'] === undefined) return false;
    if (!('hostname' in value) || value['hostname'] === undefined) return false;
    if (!('hw_address' in value) || value['hw_address'] === undefined) return false;
    if (!('ip_address' in value) || value['ip_address'] === undefined) return false;
    if (!('reserved' in value) || value['reserved'] === undefined) return false;
    if (!('service' in value) || value['service'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    if (!('subnet_id' in value) || value['subnet_id'] === undefined) return false;
    return true;
}

export function DhcpLeaseViewFromJSON(json: any): DhcpLeaseView {
    return DhcpLeaseViewFromJSONTyped(json, false);
}

export function DhcpLeaseViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): DhcpLeaseView {
    if (json == null) {
        return json;
    }
    return {
        
        'duid': json['duid'],
        'expire': json['expire'],
        'hostname': json['hostname'],
        'hw_address': json['hw-address'],
        'ip_address': json['ip-address'],
        'reserved': json['reserved'],
        'service': json['service'],
        'state': json['state'],
        'subnet_id': json['subnet-id'],
    };
}

export function DhcpLeaseViewToJSON(json: any): DhcpLeaseView {
    return DhcpLeaseViewToJSONTyped(json, false);
}

export function DhcpLeaseViewToJSONTyped(value?: DhcpLeaseView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'duid': value['duid'],
        'expire': value['expire'],
        'hostname': value['hostname'],
        'hw-address': value['hw_address'],
        'ip-address': value['ip_address'],
        'reserved': value['reserved'],
        'service': value['service'],
        'state': value['state'],
        'subnet-id': value['subnet_id'],
    };
}